Fixed some of the functionality
Realized that I should be zero-initing the host-side
y buffer then copying it to device from the get go.

Also added in the part where the X VALUE GETS SET
IN THE KERNEL.

Still doesn't work completely: it successfully calculates
for the first several y[n], then nan's the last 4 or so.